+++ Interface/AddOns/WeakAuras/Locales/frFR.lua
@@ -67,6 +67,8 @@ L["Action Usable"] = "Action utilisable"
 --[[Translation missing --]]
 L["Active"] = "Active"
 L["Additional Trigger Replacements"] = "Remplacements de Déclencheurs additionnels"
+--[[Translation missing --]]
+L["Adjust Timer"] = "Adjust Timer"
 L["Affected"] = "Affecté"
 L["Aggro"] = "Aggro"
 --[[Translation missing --]]
@@ -463,6 +465,8 @@ L["Multistrike"] = "Attaque multiple"
 L["Multi-target"] = "Multi-cibles"
 L["Music"] = "Musique"
 L["Mythic"] = "Mythique"
+--[[Translation missing --]]
+L["Mythic+ Affix"] = "Mythic+ Affix"
 L["Name"] = "Nom"
 L["Neck"] = "Cou"
 L["Neutral"] = "Neutre"