+++ Interface/AddOns/WeakAuras/Locales/itIT.lua
@@ -97,6 +97,8 @@ L["Action Usable"] = "Azione utilizzabile"
 L["Active"] = "Active"
 --[[Translation missing --]]
 L["Additional Trigger Replacements"] = "Additional Trigger Replacements"
+--[[Translation missing --]]
+L["Adjust Timer"] = "Adjust Timer"
 L["Affected"] = "Affetto"
 --[[Translation missing --]]
 L["Aggro"] = "Aggro"
@@ -677,6 +679,8 @@ L["Multi-target"] = "Multi-target"
 L["Music"] = "Music"
 --[[Translation missing --]]
 L["Mythic"] = "Mythic"
+--[[Translation missing --]]
+L["Mythic+ Affix"] = "Mythic+ Affix"
 --[[Translation missing --]]
 L["Name"] = "Name"
 --[[Translation missing --]]