+++ Interface/AddOns/WeakAuras/Locales/ptBR.lua
@@ -97,6 +97,8 @@ L["Action Usable"] = "Ação Usável"
 L["Active"] = "Active"
 --[[Translation missing --]]
 L["Additional Trigger Replacements"] = "Additional Trigger Replacements"
+--[[Translation missing --]]
+L["Adjust Timer"] = "Adjust Timer"
 L["Affected"] = "Afetado"
 --[[Translation missing --]]
 L["Aggro"] = "Aggro"
@@ -631,6 +633,8 @@ L["Multi-target"] = "Multi-alvo"
 L["Music"] = "Música"
 --[[Translation missing --]]
 L["Mythic"] = "Mythic"
+--[[Translation missing --]]
+L["Mythic+ Affix"] = "Mythic+ Affix"
 L["Name"] = "Nome"
 --[[Translation missing --]]
 L["Neck"] = "Neck"