+++ Interface/AddOns/WeakAuras/RegionTypes/Icon.lua
@@ -230,16 +230,20 @@ local function configureText(fontString, icon, enabled, point, width, height, co
     return;
   end
 
-  local sxo, syo = 0, 0;
+  local sxo, syo, h, v = 0, 0, "CENTER", "MIDDLE";
   if(point:find("LEFT")) then
     sxo = width / 10;
+    h = containment == "INSIDE" and "LEFT" or "RIGHT";
   elseif(point:find("RIGHT")) then
     sxo = width / -10;
+    h = containment == "INSIDE" and "RIGHT" or "LEFT";
   end
   if(point:find("BOTTOM")) then
     syo = height / 10;
+    v = containment == "INSIDE" and "BOTTOM" or "TOP";
   elseif(point:find("TOP")) then
     syo = height / -10;
+    v = containment == "INSIDE" and "TOP" or "BOTTOM";
   end
   fontString:ClearAllPoints();
   if(containment == "INSIDE") then
@@ -248,6 +252,8 @@ local function configureText(fontString, icon, enabled, point, width, height, co
     local selfPoint = WeakAuras.inverse_point_types[point];
     fontString:SetPoint(selfPoint, icon, point, -0.5 * sxo, -0.5 * syo);
   end
+  fontString:SetJustifyH(h);
+  fontString:SetJustifyV(v);
   local fontPath = SharedMedia:Fetch("font", font);
   fontString:SetFont(fontPath, fontSize, fontFlags == "MONOCHROME" and "OUTLINE, MONOCHROME" or fontFlags);
   fontString:SetTextHeight(fontSize);
@@ -394,17 +400,10 @@ local function modify(parent, region, data)
     local values = region.values;
     region.UpdateCustomText = function()
       WeakAuras.ActivateAuraEnvironment(region.id, region.cloneId, region.state);
-      local ok, custom = xpcall(customTextFunc, geterrorhandler(), region.expirationTime, region.duration,
-        values.progress, values.duration, values.name, values.icon, values.stacks);
-      if (not ok) then
-        custom = "";
-      end
+      values.custom = {select(2, xpcall(customTextFunc, geterrorhandler(), region.expirationTime, region.duration,
+        values.progress, values.duration, values.name, values.icon, values.stacks))}
       WeakAuras.ActivateAuraEnvironment(nil);
-      custom = WeakAuras.EnsureString(custom);
-      if(custom ~= values.custom) then
-        values.custom = custom;
-        UpdateText();
-      end
+      UpdateText();
     end
     if(data.customTextUpdate == "update") then
       WeakAuras.RegisterCustomTextUpdates(region);
@@ -467,7 +466,12 @@ local function modify(parent, region, data)
     icon:SetAllPoints();
 
     local texWidth = 1 - 0.5 * data.zoom;
-    local aspectRatio = region.keepAspectRatio and width / height or 1;
+    local aspectRatio
+    if (not region.keepAspectRatio or width == 0 or height == 0) then
+      aspectRatio = 1
+    else
+      aspectRatio = width / height;
+    end
 
     local ulx, uly, llx, lly, urx, ury, lrx, lry = GetTexCoord(region, texWidth, aspectRatio)